customized ui for plans and user show

+++ app/controllers/subscriptions_controller.rb
@@ -1,9 +1,14 @@
+# frozen_string_literal: true
+
 class SubscriptionsController < ApplicationController
   before_action :set_subscription, only: %i[destroy]
   before_action :make_payment, only: %i[create]
-
+  after_action :perform_transaction, only: %i[success]
   def create
+    authorize Subscription
+    @subscription = Subscription.new(subscription_params)
   end
+
   def success
     @subscription = Subscription.new(subscription_params)
     if @subscription.save
@@ -14,6 +19,7 @@ class SubscriptionsController < ApplicationController
   end
 
   def destroy
+    authorize Subscription
     if @subscription.destroy
       redirect_to plans_url, notice: 'Unsubscribed.'
     else
@@ -29,27 +35,20 @@ class SubscriptionsController < ApplicationController
 
   def set_subscription
     @subscription = Subscription.find(params[:id])
-  rescue ActiveRecord::RecordNotFound
-    flash[:alert] = "This subscription doesn't exist!"
-    redirect_to plans_url
   end
 
   def perform_transaction
-    Transaction.create(user_id: @subscription.user_id, subscription_id: @subscription.id, amount: @subscription.plan.monthly_fee)
+    transaction = SubscriptionService.new(@subscription).perform_transaction
+    return unless transaction
+
+    ReceiptMailer.with(user: @subscription.user, transaction: transaction).transaction_created.deliver_later
   end
 
   def make_payment
     plan = Plan.find(params[:plan_id])
-    @session = Stripe::Checkout::Session.create({
-      payment_method_types: ['card'],
-      line_items: [
-        price: plan.stripe_price_id,
-        quantity: 1,
-      ],
-      mode: 'subscription',
-      success_url: subscription_success_url(user_id: params[:user_id], plan_id: params[:plan_id]),
-      cancel_url: root_url,
-    })
-    p @session
+    redirect_to root_url, alert: 'Unable to subscribe' unless current_user.stripe_customer_id
+    success_url = subscription_success_url(user_id: params[:user_id], plan_id: params[:plan_id])
+    session = CheckoutSession.new(plan, current_user, success_url, root_url)
+    @session = session.create
   end
 end
